#5016 : Added sql validation for config payment rules + small fixes/cleanup

+++ de.metas.vertical.creditscore/creditpass/src/main/java/de/metas/vertical/creditscore/creditpass/process/validation/PaymentRuleValidation.java
@@ -42,7 +42,6 @@ public class PaymentRuleValidation extends AbstractJavaValidationRule
 			I_C_BPartner.COLUMNNAME_C_BPartner_ID);
 
 	private final CreditPassConfigRepository creditPassConfigRepository = Adempiere.getBean(CreditPassConfigRepository.class);
-	;
 
 	@Override public Set<String> getParameters()
 	{
@@ -52,7 +51,7 @@ public class PaymentRuleValidation extends AbstractJavaValidationRule
 	@Override public boolean accept(IValidationContext evalCtx, NamePair item)
 	{
 		boolean accept = false;
-		int bPartnerId = evalCtx.get_ValueAsInt(I_C_BPartner.COLUMNNAME_C_BPartner_ID, -1).intValue();
+		int bPartnerId = evalCtx.get_ValueAsInt(I_C_BPartner.COLUMNNAME_C_BPartner_ID, -1);
 
 		if (bPartnerId > -1)
 		{
